# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rythmx API
  version: 1.0.0
  description: >
    Self-hosted music automation and library management API.
    All endpoints under `/api/v1/` follow a standard envelope:
    success responses include `"status": "ok"` plus payload fields;
    error responses include `"status": "error"` and `"message"`.
    Background-triggering endpoints return HTTP 202 with `"is_alive": bool`.
servers:
  - url: /
    description: Rythmx server (all paths are absolute)

# All endpoints require an API key (X-Api-Key header) except /health
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/cruise-control/status:
    get:
      tags: [cruise-control]
      summary: Get Cruise Control pipeline run status
      operationId: getCruiseControlStatus
      responses:
        '200':
          description: Current CC pipeline status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/CruiseControlStatus'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/cruise-control/config:
    get:
      tags: [cruise-control]
      summary: Get Cruise Control configuration
      operationId: getCruiseControlConfig
      responses:
        '200':
          description: CC configuration with defaults
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/CruiseControlConfig'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [cruise-control]
      summary: Save Cruise Control configuration
      operationId: saveCruiseControlConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CruiseControlConfig'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/cruise-control/run-now:
    post:
      tags: [cruise-control]
      summary: Trigger a Cruise Control cycle immediately
      operationId: runCruiseControlNow
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/cruise-control/history:
    get:
      tags: [cruise-control]
      summary: Get Cruise Control run history
      operationId: getCruiseControlHistory
      responses:
        '200':
          description: CC run history entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      history:
                        type: array
                        items:
                          $ref: '#/components/schemas/HistoryItem'
        '500':
          $ref: '#/components/responses/ServerError'

  # ── Release Cache ─────────────────────────────────────────────────────────
  /api/v1/release-cache/clear:
    post:
      tags: [release-cache]
      summary: Clear the release cache
      operationId: clearReleaseCache
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '500':
          $ref: '#/components/responses/ServerError'

  # ── Acquisition ───────────────────────────────────────────────────────────
  /api/v1/acquisition/queue:
    get:
      tags: [acquisition]
      summary: Get acquisition queue
      operationId: getAcquisitionQueue
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AcquisitionStatus'
        - name: playlist
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Acquisition queue items
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      queue:
                        type: array
                        items:
                          $ref: '#/components/schemas/QueueItem'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [acquisition]
      summary: Add an item to the acquisition queue
      operationId: addAcquisitionItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [artist, album, kind]
              properties:
                artist:
                  type: string
                album:
                  type: string
                kind:
                  $ref: '#/components/schemas/ReleaseKind'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/acquisition/stats:
    get:
      tags: [acquisition]
      summary: Get acquisition queue statistics
      operationId: getAcquisitionStats
      responses:
        '200':
          description: Queue counts by status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/AcquisitionStats'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/acquisition/check-now:
    post:
      tags: [acquisition]
      summary: Trigger the acquisition worker immediately
      operationId: checkAcquisitionNow
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'

  # ── Images ────────────────────────────────────────────────────────────────
  /api/v1/images/resolve:
    post:
      tags: [images]
      summary: Resolve image URL for a music entity
      operationId: resolveImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, name]
              properties:
                type:
                  $ref: '#/components/schemas/ImageType'
                name:
                  type: string
                artist:
                  type: string
      responses:
        '200':
          description: Resolved or pending image URL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/ImageResolveResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/images/warm-cache:
    post:
      tags: [images]
      summary: Proactively warm the image cache
      operationId: warmImageCache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [type, name]
                    properties:
                      type:
                        $ref: '#/components/schemas/ImageType'
                      name:
                        type: string
                      artist:
                        type: string
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'

  # ── Playlists ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >
        Auto-generated API key. View or regenerate in Settings.
        Included as `X-Api-Key: <key>` on every request.
...
  responses:
...
    Ok:
      description: Operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiSuccess'
    Accepted:
      description: Background operation triggered
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiSuccess'
              - type: object
                required: [is_alive]
                properties:
                  is_alive:
                    type: boolean
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Conflict:
      description: Operation already in progress or resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    Period:
      type: string
      enum: [7day, 1month, 3month, 6month, 12month, overall]
    AcquisitionStatus:
      type: string
      enum: [pending, submitted, found, failed, skipped]
    ReleaseKind:
      type: string
      enum: [album, ep, single, compilation]
...
    ImageType:
      type: string
      enum: [artist, album, track]

    # ── Envelope ────────────────────────────────────────────────────────────
    ApiSuccess:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ApiError:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        error:
          type: string

    # ── Music Entities ───────────────────────────────────────────────────────
...
    QueueItem:
      type: object
      required: [id, artist, album, kind, status]
      properties:
        id:
          oneOf:
            - type: string
            - type: integer
        artist:
          type: string
        album:
          type: string
        kind:
          $ref: '#/components/schemas/ReleaseKind'
        status:
          $ref: '#/components/schemas/AcquisitionStatus'
        requested_by:
          type: string
        requested_at:
          type: string
          format: date-time
        release_date:
          type: string
    AcquisitionStats:
      type: object
      required: [pending, submitted, found, failed, skipped, total]
      properties:
        pending:
          type: integer
        submitted:
          type: integer
        found:
          type: integer
        failed:
          type: integer
        skipped:
          type: integer
        total:
          type: integer

    # ── Playlists ────────────────────────────────────────────────────────────
...
    CruiseControlConfig:
      type: object
      properties:
        run_mode:
          type: string
          enum: [build, fetch]
        playlist_prefix:
          type: string
        min_listens:
          type: integer
        period:
          $ref: '#/components/schemas/Period'
        lookback_days:
          type: integer
        max_per_cycle:
          type: integer
        max_playlist_tracks:
          type: integer
        auto_push_playlist:
          type: boolean
        schedule_weekday:
          type: integer
          minimum: 0
          maximum: 6
        schedule_hour:
          type: integer
          minimum: 0
          maximum: 23
        enabled:
          type: boolean
        dry_run:
          type: boolean
        release_cache_refresh_weekday:
          type: integer
          minimum: 0
          maximum: 6
        release_cache_refresh_hour:
          type: integer
          minimum: 0
          maximum: 23
        nr_ignore_keywords:
          type: string
        nr_ignore_artists:
          type: string
        include_features:
          type: boolean
        release_kinds:
          type: string
          description: Comma-separated list of ReleaseKind values to include
    CruiseControlStatus:
      type: object
      required: [state]
      properties:
        state:
          type: string
          enum: [idle, running, error, completed]
        stage:
          type: integer
        stage_label:
          type: string
        total_stages:
          type: integer
        last_run:
          type: string
          format: date-time
        summary:
          type: object
          properties:
            artists_checked:
              type: integer
            new_releases:
              type: integer
            owned:
              type: integer
            queued:
              type: integer
        error:
          type: string
    HistoryItem:
      type: object
      required: [artist, album, status, date]
      properties:
        artist:
          type: string
        album:
          type: string
        status:
          type: string
          enum: [owned, queued, skipped]
        reason:
          type: string
        date:
          type: string
          format: date-time

    # ── Images ───────────────────────────────────────────────────────────────
    ImageResolveResponse:
      type: object
      properties:
        image_url:
          type: string
        pending:
          type: boolean

    # ── Stats ────────────────────────────────────────────────────────────────
...